rename fixtures and test specs
