+++ fixtures/chain-fixtures.fixture.ts
@@ -1,6 +1,6 @@
 import { mergeTests } from '@playwright/test';
-import { pageObjectTest } from './PageObjects.fixture';
-import { sauceTest } from './saucedemoTestData.fixture';
+import { pageObjectTest } from './page-objects.fixture';
+import { sauceTest } from './saucedemo-test-data.fixture';
 
 export const test = mergeTests(pageObjectTest, sauceTest);
 export { expect } from '@playwright/test';
