+++ fixtures/restfulbooker-api-objects.fixture.ts
