+++ fixtures/springboot-api-objects.fixture.ts
