+++ tests/ui/saucedemo/Login.spec.ts
@@ -1,4 +1,4 @@
-import { test } from '../../../fixtures/chainFixtures';
+import { test } from '../../../fixtures/chain-fixtures.fixture';
 
 test.describe
 	.skip('Unsuccessful Login Scenarios', () => {
