+++ tests/ui/saucedemo/Shopping.spec.ts
@@ -1,4 +1,4 @@
-import { test } from '../../../fixtures/chainFixtures';
+import { test } from '../../../fixtures/chain-fixtures.fixture';
 
 test.describe('Shopping Scenarios', () => {
 	test('完整流程從購物到結帳', async ({
